+++ CHANGES.md
@@ -1,6 +1,7 @@
 0.1.2
 ----
 - Disable printing to stderr because OpenSSH never pushes `pam_info` messages until an after-login banner anyway
+- Hide debug logs from reqwest/hyper
 
 0.1.1
 -----